requirements.txt
altgraph==0.17
asgiref==3.2.7
beautifulsoup4==4.9.1
bottle==0.12.18
bottle-websocket==0.2.9
cefpython3==66.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.9
django-bootstrap4-datetimepicker==4.2
django-channels==0.7.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-filter==2.1.0
django-htmlmin==0.11.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-scheduler==0.9.3
django-sitetree==1.13.1
django-taggit==1.1.0
Eel==0.12.4
et-xmlfile==1.0.1
flake8==3.8.3
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
html5lib==1.1
icalendar==4.0.4
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
langdetect==1.0.8
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==1.4.2.post1
nexmo==2.4.0
numpy==1.18.5
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
pandas==1.0.5
pefile==2019.4.18
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyInstaller==3.6
PyJWT==1.7.1
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.1.2
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==0.14.0
unicodecsv==0.14.1
urllib3==1.25.9
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==4.1.2
x-django-app==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
